pyproject.toml
[project]
name = "agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Muhammad Toqeer", email = "166752233+t0q33r@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.1",
    "openai-agents>=0.0.13",
    "plotly>=6.0.1",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.44.1",
]

[project.scripts]
agent-project = "agent_project:main"
agent2 = "agent_project.main:agent_func"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
